refactor(core/message): decompose
